requirements.txt
python-dotenv
langgraph
langchain
langchain_core
langchain_community
langchain_google_genai
langchain_text_splitters
chromadb
google-genai
pandas
gradio
typing-extensions
protobuf
unstructured
unstructured[pdf]